create vegetation mod
