+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod contours;
 pub mod knolls;
 pub mod merge;
 pub mod util;
+pub mod vegetation;
